+++ src/components/Sidebar/FriendsList/component.tsx
@@ -4,31 +4,30 @@ import { FC, useState } from "react";
 import { FriendsListProps } from "./interface";
 import axios from "axios";
 
-const FriendsList: FC<FriendsListProps> = (props) => {
-    const { friends } = props;
+const FriendsList: FC<FriendsListProps> = ({ friends = [] }) => {
+    console.log("friends:")
+    console.log(friends);
     const [isVisible, setIsVisible] = useState(false);
-
     const openPopup = () => setIsVisible(true);
     const closePopup = () => setIsVisible(false);
-
     const apiPost = async (senderId: string) => {
-        await axios.post('/api/friends/remove', { id: senderId });
+        await axios.post('/api/friends/remove', { idToRemove: senderId });
     }
 
+
+    const sortedFriends = friends.sort((friendA, friendB) => friendA.name?.localeCompare(friendB.name));
     return (
         <ul className='sidebar-chat-list'>
-            {friends?.sort((friendA: string, friendB: string) => {
-                return friendA > friendB ? 1 : -1
-            }).map((friend: string) => {
-                return (<div><li className='sidebar-chat-list-item'>
-                    <span className='truncate' onClick={!isVisible ? openPopup : closePopup}>{friend}</span>
+            {sortedFriends.map((friend) => {
+                return (<div key={friend.id}><li className='sidebar-chat-list-item'>
+                    <span className='truncate' onClick={!isVisible ? openPopup : closePopup}>{friend.name}</span>
                 </li>
                     {isVisible && (
                         <li className='sidebar-chat-list-item'>
                             <div className="flex flex-col">
                                 <ul className="flex flex-col pl-4">
                                     <li>Chat</li>
-                                    <li onClick={() => apiPost('stub')}>Remove Friend</li>
+                                    <li onClick={() => apiPost(friend.id)}>Remove Friend</li>
                                 </ul>
                             </div>
                         </li>
